refactor(tui): extract passphrase generator from serve dialog
Cleanup-pass finding #3 from the code-quality scan. Partial split.

src/tui/dialogs/serve.rs (2,554 lines) is now src/tui/dialogs/serve/
with mod.rs (2,394 lines) and passphrase.rs (171 lines). The full
phase-by-phase split the scan suggested (mode_picker / confirm /
active / log_tail submodules) would require many pub(super) annotations
to thread state across the state machine and was judged a poor cost-
benefit on careful review — the phases share too much mutable state
on `ServeDialog` to cleanly tear apart.

The passphrase generator and 96-word list are genuinely self-contained
(only the public `generate_passphrase` is consumed) and move cleanly
with their three tests. mod.rs imports `passphrase::generate_passphrase`.
